refactor: pull nuke utilities into subfile
